boot: Add exception level 2 system control registers initialization
This is another step in change of a CPU boot implementation to allow
a software to use EL3, EL1 and EL0 exception levels.

This commit adds minimal configuration to system control registers
related with exception level 2 (EL2). That allows to switch to
exception level 1 (EL1).

The EL2 is minimal, just to provide a configuration required to
return to EL1, hence there is no:
- trapping of access to any features from EL1, EL0
- trapping of usage of any instructions from EL1, EL0,
- virtualization,
- HVC instruction is disabled.

The EL2 layer is transparent. There is no plan to implement
support for virtualization in Standalone BSP.

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>